utils/pressure_test/requirements.txt
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
ConfigArgParse==1.5.1
cycler==0.10.0
cypari==2.4.1
decorator==5.0.9
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
future==0.18.2
FXrays==1.3.5
gevent==21.1.2
geventhttpclient==1.4.5
greenlet==1.1.0
idna==2.10
importlib-metadata==4.6.1
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
kiwisolver==1.3.1
knot-floer-homology==1.0.2
locust==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
msgpack==1.0.2
networkx==2.6.1
numpy==1.21.0
pandas==1.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plink==2.4
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
pypng==0.0.20
python-dateutil==2.8.1
python-snappy==0.6.0
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
scipy==1.7.0
six==1.16.0
snappy-manifolds==1.1.2
spherogram==2.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
